BUG-876 fix race condition
The page displayed an error message if valintakokeet promise
was resolved after valintakoetulokset got resolved.

Fix by ensuring that both promises have been resolved before
processing the data.

This file is in a need of a big rewrite, but it is out
of the scope of the expedite bug fix.